updated individual service ui & migrated get apis for review

+++ lib/action/review.ts
@@ -2,6 +2,7 @@
 
 import { prisma } from "@/lib/prisma";
 import { Review } from "@/types/review";
+import { revalidatePath } from "next/cache";
 
 // Create a review
 export async function createReview(
@@ -10,7 +11,14 @@ export async function createReview(
   try {
     const review = await prisma.review.create({
       data,
+      include: {
+        service: {
+          select: { slug: true },
+        },
+      },
     });
+
+    revalidatePath(`/service/${review.service.slug}`);
     return review;
   } catch (error) {
     console.log("Error creating review:", error);
@@ -31,28 +39,25 @@ export async function getReviewsByService(serviceId: string) {
   }
 }
 
-export async function getReviewById(id: string) {
-  try {
-    const review = await prisma.review.findUnique({
-      where: { id },
-    });
-    return review;
-  } catch (error) {
-    console.log("Error fetching review by ID:", error);
-    return null;
-  }
-}
-
 // Update a review
 export async function updateReview(
   id: string,
   data: Partial<Omit<Review, "id" | "createdAt" | "updatedAt">>,
 ) {
   try {
-    return prisma.review.update({
+    const review = await prisma.review.update({
       where: { id },
       data,
+      include: {
+        service: {
+          select: { slug: true },
+        },
+      },
     });
+
+    revalidatePath(`/service/${review.service.slug}`);
+
+    return review;
   } catch (error) {
     console.log("Error updating review:", error);
     return null;
@@ -62,34 +67,25 @@ export async function updateReview(
 // Delete a review
 export async function deleteReview(id: string) {
   try {
-    return prisma.review.delete({
+    const review = await prisma.review.findUnique({
+      where: { id },
+      include: {
+        service: {
+          select: { slug: true },
+        },
+      },
+    });
+
+    if (!review) return null;
+
+    await prisma.review.delete({
       where: { id },
     });
+
+    revalidatePath(`/service/${review.service.slug}`);
+    return review;
   } catch (error) {
     console.log("Error deleting review:", error);
     return null;
-  }
-}
-
-export async function getAllReviewsGroupedByService() {
-  try {
-    const services = await prisma.services.findMany({
-      where: {
-        reviews: {
-          some: {},
-        },
-      },
-      include: {
-        reviews: {
-          orderBy: { createdAt: "desc" },
-        },
-      },
-      orderBy: { name: "asc" },
-    });
-
-    return services;
-  } catch (error) {
-    console.error("Error fetching reviews:", error);
-    return [];
   }
 }
